# PERF: resolution, is_normalized